week_5/requirements.txt
albumentations==0.4.5
opencv_python_headless==4.2.0.34
motmetrics==1.2.0
scikit_image==0.15.0
numpy==1.16.4
tqdm==4.41.1
matplotlib==3.1.3
scipy==1.4.1
image_classifiers==1.0.0
filterpy==1.4.5
Keras==2.2.4
keras_rectified_adam==0.17.0
classification_models==0.1
imageio==2.8.0
imageio-ffmpeg==0.4.1
keras_radam==0.15.0
Pillow==7.1.1
plotly==4.6.0
PyYAML==5.3.1
scikit_learn==0.22.2.post1
